+++ WorldCities/WorldCities/src/app/app.module.ts
@@ -9,14 +9,16 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HomeComponent } from './home/home.component';
 import { NavMenuComponent } from './nav-menu/nav-menu.component';
-import { CitiesComponent } from './cities/cities.component'
+import { CitiesComponent } from './cities/cities.component';
+import { CountriesComponent } from './countries/countries.component'
 
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
     NavMenuComponent,
-    CitiesComponent
+    CitiesComponent,
+    CountriesComponent
   ],
   imports: [
     BrowserModule,