+++ WorldCities/WorldCities/src/app/countries/countries.component.ts
@@ -1,0 +1,68 @@
+import { HttpClient, HttpParams } from '@angular/common/http';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatPaginator, PageEvent } from '@angular/material/paginator';
+import { MatSort } from '@angular/material/sort';
+import { MatTableDataSource } from '@angular/material/table';
+import { environment } from '../../envrionments/environment';
+import { Country } from './country';
+
+@Component({
+  selector: 'app-countries',
+  templateUrl: './countries.component.html',
+  styleUrl: './countries.component.scss'
+})
+export class CountriesComponent implements OnInit {
+  public displayedColumns: string[] = ['id', 'name', 'iso2', 'iso3'];
+
+  public countries!: MatTableDataSource<Country>;
+
+  defaultPageIndex: number = 0;
+  defaultPageSize: number = 10;
+  defaultFilterColumn: string = "name";
+  filterQuery?: string;
+  public defaultSortColumn: string = "id";
+  public defaultSortOrder: "asc" | "desc" = "asc";
+
+  @ViewChild(MatPaginator) paginator!: MatPaginator;
+  @ViewChild(MatSort) sort!: MatSort;
+
+  constructor(private http: HttpClient) { }
+
+  ngOnInit() {
+    this.loadData();
+  }
+
+  loadData(query?: string) {
+    var pageEvent = new PageEvent(); // generated a page event to call getData for the initial component load
+    pageEvent.pageIndex = this.defaultPageIndex;
+    pageEvent.pageSize = this.defaultPageSize;
+    this.filterQuery = query;
+    console.log(query);
+    this.getData(pageEvent);
+  }
+
+  // Used by mat-paginator element and the ngOnInit()
+  getData(event: PageEvent) {
+    var url = environment.baseURL + 'api/Countries';
+    var params = new HttpParams()
+      .set("pageIndex", event.pageIndex.toString())
+      .set("pageSize", event.pageSize.toString())
+      .set("sortColumn", (this.sort) ? this.sort.active : this.defaultSortColumn)
+      .set("sortOrder", (this.sort) ? this.sort.direction : this.defaultSortOrder);
+
+    if (this.filterQuery) {
+      params = params
+        .set("filterColumn", this.defaultFilterColumn)
+        .set("filterQuery", this.filterQuery)
+    }
+
+    this.http.get<any>(url, { params })
+      .subscribe(result => {
+        this.paginator.length = result.totalCount;
+        this.paginator.pageIndex = result.pageIndex;
+        this.paginator.pageSize = result.pageSize;
+        this.countries = new MatTableDataSource<Country>(result.data);
+      }, error => console.error(error));
+  }
+
+}