+++ WorldCities/WorldCitiesAPI/Data/Models/Country.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json.Serialization;
 
 namespace WorldCitiesAPI.Data.Models
 {
@@ -22,15 +23,17 @@ namespace WorldCitiesAPI.Data.Models
     /// Country Name (in UTF8 format)
     /// </summary>
     public string Name { get; set; } = null!;
-    
+
     /// <summary>
     /// Country code (in ISO 3166-1 ALPHA-2 format)
     /// </summary>
+    [JsonPropertyName("iso2")]
     public string ISO2 { get; set; } = null!;
 
     /// <summary>
     /// Country code (in ISo 3166-1 ALPHA-3 format)
     /// </summary>
+    [JsonPropertyName("iso3")]
     public string ISO3 { get; set; } = null!;
     #endregion