+++ WorldCities/WorldCitiesAPI/CHANGELOG.md
@@ -42,3 +42,5 @@ The following steps were used to generate this project:
 - Create new `Data/ApiResult.cs` file with pageIndex and pageSize
 - Modify `ApiResult.cs` add sorting functionality
 - Modify `CitiesController.cs` update GetCities with sorting functionality.
+- Modify `Data/Country.cs` add JsonPropertyName("iso3") for iso2 and iso3
+-